Made expression preview tabs initially hidden
The display:none style was removed from the tab content elements. This
was needed earlier since the tab contents would otherwise not show.
Recent changes seems to have fixed the original issue and the previous
fix now makes the contents of all tabs show when opening the modal.

Removing the lines which removed the display:none style fixes this
issue.